Move request method to alertv2 client.
